tests/ux/spatnav-distance-function-grid-001.html

<!doctype html>
<meta charset="utf-8">
<title>Spatnav sanity check</title>
<link rel="author" title="Jihye Hong" href="jh.hong@lge.com">
<meta name=flags content="">
<link rel=help href="https://wicg.github.io/spatial-navigation/#high-level-api">

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--load polyfill. To be removed before submitting to WPT-->
<script src="../polyfill/spatnav-heuristic.js"></script>

<style>
  :focus {
    outline: 4px #3582ff solid;
  }

  .container {
    margin: 10px;
  }

  .box {
    position: relative;
    box-sizing: border-box;
    width: 50px;
    height: 50px;
    border: 1px black solid;
    background: #6ac;
  }

  #initial_focus {
    background: #F08080;
  }
</style>

<div class="container">
  <div tabindex="0" class="box"></div>
  <div tabindex="0" class="box" style="left: 50px; top: -50px;"></div>
  <div tabindex="0" class="box" id="up_focus" style="left: calc(50px * 2); top: calc(-50px * 2);"></div>
  <div tabindex="0" class="box" style="top: calc(-50px * 2);"></div>
  <div tabindex="0" class="box" id="left_focus" style="left: 50px; top: calc(-50px * 3);"></div>
  <div tabindex="0" class="box" id="initial_focus" style="left: calc(50px * 2); top: calc(-50px * 4);"></div>
  <div tabindex="0" class="box" style="top: calc(-50px * 4);"></div>
  <div tabindex="0" class="box" style="left: 50px; top: calc(-50px * 5);"></div>
  <div tabindex="0" class="box" id="down_focus" style="left: calc(50px * 2); top: calc(-50px * 6);"></div>
</div>

<script>
test(function() {
  document.getElementById("initial_focus").focus();
  window.navigate("left");
  assert_equals(document.activeElement.id, "left_focus");
}, "Move focus left from the initially focused element")

test(function() {
  document.getElementById("initial_focus").focus();
  window.navigate("up");
  assert_equals(document.activeElement.id, "up_focus");
}, "Move focus up from the initially focused element")

test(function() {
  document.getElementById("initial_focus").focus();
  window.navigate("down");
  assert_equals(document.activeElement.id, "down_focus");
}, "Move focus down from the initially focused element")

</script>
